+++ src/Entities/AbstractEntityCollection.php
@@ -6,6 +6,7 @@ use Countable;
 use Doctrine\Common\Collections\Collection;
 use IteratorAggregate;
 use Traversable;
+use function max;
 
 /**
  * Class AbstractEntityCollection
@@ -13,6 +14,10 @@ use Traversable;
  * Provides a wrapper for managing child entities outside of an aggregate root.
  * This allows logic to still be tied with the aggregate, but moved out to a sub-object
  * to keep the main aggregate logic smaller.
+ *
+ * This class is intended to be used with the AbstractEntity class that uses integers for
+ * the object identities. If you require a different scheme e.g.: some custom string then
+ * implement your own base logic for those use-cases.
  *
  * When implementing be sure to add additional methods as necessary to provide the domain
  * implementations that are needed.
@@ -25,11 +30,13 @@ abstract class AbstractEntityCollection implements Countable, IteratorAggregate
 
     protected AggregateRoot $root;
     protected Collection $entities;
+    protected int $lastId;
 
     public function __construct(AggregateRoot $root, Collection $entities)
     {
         $this->root     = $root;
         $this->entities = $entities;
+        $this->lastId   = $this->findLastId();
     }
 
     public function getIterator(): Traversable
@@ -54,10 +61,11 @@ abstract class AbstractEntityCollection implements Countable, IteratorAggregate
 
     protected function nextId(): int
     {
-        if (0 === $this->count()) {
-            return 1;
-        }
+        return ++$this->lastId;
+    }
 
-        return (int)max($this->entities->map(fn (AbstractEntity $e) => $e->id())->getValues()) + 1;
+    protected function findLastId(): int
+    {
+        return $this->count() > 0 ? (int)max($this->entities->map(fn (AbstractEntity $e) => $e->id())->getValues()) : 0;
     }
 }